Move dalvik opcode.h into anal/arch
