+++ libr/anal/p/dalvik.mk
@@ -6,4 +6,4 @@ TARGET_DALVIK=anal_dalvik.${EXT_SO}
 ALL_TARGETS+=${TARGET_DALVIK}
 
 ${TARGET_DALVIK}: ${OBJ_DALVIK}
-	${CC} $(call libname,anal_dalvik) ${LDFLAGS} ${CFLAGS} -I../../asm/arch -o anal_dalvik.${EXT_SO} ${OBJ_DALVIK}
+	${CC} $(call libname,anal_dalvik) ${LDFLAGS} ${CFLAGS} -o anal_dalvik.${EXT_SO} ${OBJ_DALVIK}
